Interim step towards multiple renderers
