--- lib/msfl_visitors/renderers/render_constant.rb
@@ -1,12 +1,0 @@
-module MSFLVisitors
-  module Renderers
-    class RenderConstant < Renderer
-      def initialize(constant)
-        @rendered = constant
-      end
-      def render
-        @rendered
-      end
-    end
-  end
-end
