--- lib/msfl_visitors/renderers/render_string_value.rb
@@ -1,10 +1,0 @@
-module MSFLVisitors
-  module Renderers
-    class RenderStringValue < Renderer
-      def render
-        node.value.to_s
-      end
-    end
-  end
-end
-
