--- lib/msfl_visitors/renderers/render_temporal.rb
@@ -1,9 +1,0 @@
-module MSFLVisitors
-  module Renderers
-    class RenderTemporal < Renderer
-      def render
-        node.value.iso8601
-      end
-    end
-  end
-end
