--- lib/msfl_visitors/renderers/render_value.rb
@@ -1,9 +1,0 @@
-module MSFLVisitors
-  module Renderers
-    class RenderValue < Renderer
-      def render
-        node.value
-      end
-    end
-  end
-end
