--- lib/msfl_visitors/renderers/renderer.rb
@@ -1,13 +1,0 @@
-module MSFLVisitors
-  module Renderers
-    class Renderer
-      def initialize(node)
-        @node = node
-      end
-
-      private
-
-      attr_reader :node
-    end
-  end
-end
