+++ lib/msfl_visitors/visitors/base.rb
@@ -1,15 +1,19 @@
 module MSFLVisitors
   module Visitors
     class Base
-      attr_accessor :collector
 
-      def initialize(collector)
-        self.collector = collector
+      def initialize(collector, renderer=Renderers::ChewyRenderer)
+        @collector = collector
+        @renderer = renderer.new
       end
 
       def visit(obj)
         collector << render(obj)
       end
+
+      private
+
+      attr_reader :collector
     end
   end
 end
